Remove separate student and admin presigned post url routes
